organized main section components into main-section
