+++ client/src/components/overview/main-section/productinfo.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import styled from 'styled-components';
+
+function ProductInfo() {
+  return <Info />
+}
+
+const Info = styled.div`
+  background-color: #CCC;
+`;
+
+export default ProductInfo;
