--- client/src/components/overview/productdescription.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-function ProductDescription() {
-  return <Description />
-}
-
-const Description = styled.div`
-  background-color: #CCC;
-`;
-
-export default ProductDescription;
