+++ client/src/components/overview/title-section/titlebar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import logo from '../img/logo.jpg';
-import SearchBar from '../searchbar.jsx';
+import SearchBar from './searchbar.jsx';
 
 function TitleBar() {
   return (
